update subject list

+++ plotting/plot_spectra.m
@@ -15,7 +15,7 @@ function plot_spectra(EEG, condition, subject, result_dir, varargin)
 
     g = finputcheck(varargin, { ...
         'title_str'      'string'        []              [];
-        'f1'             'integer'       { }             10;
+        'f1'             'double'       { }             10;
         });
     if ischar(g), error(g); end
